adaptive partial aggregation: fix bug in agg filter, fix comment

+++ core/src/type/decimal_operations.h
@@ -428,7 +428,7 @@ public:
             return;
         }
         if (value > 0) {
-            val = static_cast<uint128_t>(value);
+            val = value;
             signum = 1;
         } else {
             int128_t tmp = value;
